lab_02/img_to_poly/requirements.txt
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
cffi==1.15.1
defusedxml==0.7.1
executing==1.2.0
fastjsonschema==2.16.3
fqdn==1.5.1
idna==3.4
imageio==2.26.0
importlib-metadata==6.0.0
isoduration==20.11.0
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
lazy_loader==0.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.20.0
scipy==1.9.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
svgwrite==1.4.3
terminado==0.17.1
tifffile==2023.3.15
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
uri-template==1.2.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
zipp==3.15.0
